% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfast-deprecated.R, R/create16dayts.R
\name{bfast-deprecated}
\alias{bfast-deprecated}
\alias{create16dayts}
\title{Deprecated functions in package \pkg{bfast}.}
\usage{
create16dayts(data, dates)
}
\description{
The functions listed below are deprecated and will be removed in the future.
Please use the alternative functions as listed below.
Help pages for deprecated functions are available via \code{help("<function>-deprecated")}.
}
\section{\code{create16dayts}}{

Use \code{\link{bfastts}} with the parameter \code{type="16-day"}.
}

\keyword{internal}
