% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create16dayts.R
\name{create16dayts-deprecated}
\alias{create16dayts-deprecated}
\title{A helper function to create time series}
\usage{
create16dayts(data, dates)
}
\arguments{
\item{data}{Passed to bfastts.}

\item{dates}{Passed to bfastts.}
}
\description{
A deprecated alias to bfastts.
Please use \code{bfastts(type="16-day")} instead.
}
\seealso{
\code{\link[bfast]{bfastmonitor}}

\code{\link{bfast-deprecated}}
}
\author{
Achim Zeileis, Jan Verbesselt
}
\keyword{ts}
