% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create16dayts.R
\name{create16dayts}
\alias{create16dayts}
\title{A helper function to create time series}
\usage{
create16dayts(data, dates)
}
\arguments{
\item{data}{A vector}

\item{dates}{A vector ....}
}
\description{
Time series creation
}
\examples{

## set up a 16-day time series
#ndvi <- create16dayts(modisraster[1],dates)
#plot(ndvi)

}
\seealso{
\code{\link[bfast]{bfastmonitor}}
}
\author{
Achim Zeileis, Jan Verbesselt
}
\keyword{ts}
