% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfast-package.R
\docType{data}
\name{dates}
\alias{dates}
\title{A vector with date information (a Datum type) to be linked with each NDVI
layer within the modis raster datacube (modisraster data set)}
\source{
\insertRef{janbfastmonitor}{bfast}
}
\description{
\code{dates} is an object of class "Date" and contains the "Date"
information to create a 16-day time series object.
}
\examples{

## see ?bfastmonitor for examples

}
\keyword{datasets}
