% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.bfast_cpp_closestfrom}
\alias{.bfast_cpp_closestfrom}
\title{For all elements of a vector a, find the closest elements in a vector B and returns resulting indexes}
\usage{
.bfast_cpp_closestfrom(a, b, twosided)
}
\arguments{
\item{a}{numeric vector, ordered}

\item{b}{numeric vector, ordered}

\item{twosided}{logical value, if false, indexes will always point to elements in b that are less than or equal to elements in a but not greater than.}
}
\value{
integer vector of the same size as a with elements represnting indexes pointing to closest values in b
}
\description{
For all elements of a vector a, find the closest elements in a vector B and returns resulting indexes
}
