% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFAST.R, R/RcppExports.R
\name{ICMEM}
\alias{ICMEM}
\title{ICMEM.}
\usage{
ICMEM(
  Y,
  x_int,
  Adj,
  A_int,
  mu0_int,
  W_int,
  mu_int,
  sigma_int,
  lambda_grid,
  alpha,
  beta_grid,
  maxIter_ICM,
  maxIter
)

ICMEM(
  Y,
  x_int,
  Adj,
  A_int,
  mu0_int,
  W_int,
  mu_int,
  sigma_int,
  lambda_grid,
  alpha,
  beta_grid,
  maxIter_ICM,
  maxIter
)
}
\arguments{
\item{Y}{is a N*p feature matrix}

\item{x_int}{is a vector of initial cluster label.}

\item{Adj}{is a matrix containing neighborhood information generated by getneighborhood_fast.}

\item{A_int}{is an initial p*D factor loading matrix.}

\item{mu0_int}{is an initial p*1 mean vector.}

\item{W_int}{is an initial p*p diagonal matrix.}

\item{mu_int}{is an initial mean vector. we often generated it by Gaussian mixture model.}

\item{sigma_int}{is a initial co-variance matrix. we often generated it by Gaussian mixture model.}

\item{lambda_grid}{is a sequence of exponential decay parameter in the zero-inflation model.}

\item{alpha}{is a intercept.}

\item{beta_grid}{is a sequence of smoothing parameter that can be specified by user.}

\item{maxIter_ICM}{is the maximum iteration of ICM algorithm.}

\item{maxIter}{is the maximum iteration of EM algorithm.}

\item{y}{is a matrix of PCs containing gene expression.}

\item{PX}{is a logical value specifying the parameter expansion in EM algorithm.}

\item{theta_int}{is a initial precision matrix. we often generated it by Gaussian mixture model.}

\item{maxiter_ICM}{is the maximum iteration of ICM algorithm.}

\item{maxiter}{is the maximum iteration of EM algorithm.}
}
\value{
a list.

The item 'x' is the clustering result.

The item 'gam' is the posterior probability matrix.

The item 'ell' is the opposite log-likelihood.

The item 'mu' is the mean of each component.

The item 'sigma' is the variance of each component.

a list.
The item 'x' is the clustering result.
The item 'gam' is the posterior probability matrix.
The item 'A' is a factor lodading matrix.
The item 'mu0' is the mean of the residual.
The item 'W' is the variance of the residual.
The item 'mu' is the mean of each component.
The item 'sigma' is the variance of each component.
}
\description{
The function ICMEM was used to conduct spatial clustering with hidden Markov random field for a sequence of beta and fixed number of clusters

estimates the joint latent expectations in the E-step.
}
\details{
The function ICMEM was used to conduct spatial clustering with hidden Markov random field for fixed beta and fixed number of clusters
}
