% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfast-package.R
\docType{data}
\name{modisraster}
\alias{modisraster}
\title{A raster brick of 16-day satellite image NDVI time series for a small subset
in south eastern Somalia.}
\source{
Verbesselt, J., R. Hyndman, G. Newnham, and D. Culvenor (2012).
Near real-time disturbance detection using satellite image time series.
\emph{Remote Sensing of Environment}.
\url{http://eeecon.uibk.ac.at/wopec2/repec/inn/wpaper/2011-18.pdf}.
}
\description{
A raster brick containing 16-day NDVI satellite images (MOD13C1 product).
}
\examples{


## see ?bfastmonitor


}
\keyword{datasets}
\keyword{ts}
