% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfast-package.R
\docType{data}
\name{ndvi}
\alias{ndvi}
\title{A random NDVI time series}
\description{
A univariate time series object of class "ts". Frequency is set to 24.
}
\examples{

plot(ndvi)

}
\keyword{datasets}
\keyword{ts}
