% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfast-package.R
\docType{data}
\name{som}
\alias{som}
\title{Two 16-day NDVI time series from the south of Somalia}
\source{
\insertRef{janbfastmonitor}{bfast}
}
\description{
\code{som} is a dataframe containing time and two NDVI time series to
illlustrate how the monitoring approach works.
}
\examples{

## first define the data as a regular time series (i.e. ts object)
library(zoo)
NDVI <- as.ts(zoo(som$NDVI.b,som$Time))
plot(NDVI)


}
\keyword{datasets}
