% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wpcaCpp}
\alias{wpcaCpp}
\title{wpcaCpp}
\usage{
wpcaCpp(Y, nPCs, weighted = TRUE)
}
\arguments{
\item{Y}{is the observed n*p feature matrix.}

\item{nPCs}{is the number of the principle components.}
}
\description{
an efficient function to find the
}
