\name{dates}
\alias{dates}
\docType{data}
\title{
A vector with date information (a Datum type) to be linked with each NDVI layer within the modis raster brick (modisraster data set)
}
\description{
\code{dates} is an object of class "Date" and contains the "Date" information to create a 16-day time series object.
}


\source{
Verbesselt, J., R. Hyndman, G. Newnham, and D. Culvenor (2012). 
Near real-time disturbance detection using satellite image time series. 
\emph{Remote Sensing of Environment}. \url{http://eeecon.uibk.ac.at/wopec2/repec/inn/wpaper/2011-18.pdf}.
}

\examples{
## see ?bfastmonitor for examples
}
\keyword{datasets}
