% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfast-package.R
\docType{data}
\name{harvest}
\alias{harvest}
\title{16-day NDVI time series for a Pinus radiata plantation.}
\source{
\insertRef{janbfast}{bfast}
}
\description{
A univariate time series object of class "ts". Frequency is set to 23 -- the
approximate number of observations per year.
}
\examples{

plot(harvest,ylab='NDVI')

}
\keyword{datasets}
\keyword{ts}
