% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfast-package.R
\docType{data}
\name{modisraster}
\alias{modisraster}
\title{A raster datacube of 16-day satellite image NDVI time series for a small subset
in south eastern Somalia.}
\source{
\insertRef{janbfastmonitor}{bfast}
}
\description{
A Cloud-Optimised GeoTIFF containing 16-day NDVI satellite images (MOD13C1 product).
}
\examples{


## see ?bfastmonitor


}
\keyword{datasets}
\keyword{ts}
