\name{modisraster}
\alias{modisraster}
\docType{data}
\title{
A raster brick of 16-day satellite image NDVI time series for a small subset in south eastern Somalia.
}
\description{
A raster brick containing 16-day NDVI satellite images (MOD13C1 product).
}

\source{
Verbesselt, J., R. Hyndman, G. Newnham, and D. Culvenor (2012). 
Near real-time disturbance detection using satellite image time series. 
\emph{Remote Sensing of Environment}. \url{http://eeecon.uibk.ac.at/wopec2/repec/inn/wpaper/2011-18.pdf}.
}

\examples{

## see ?bfastmonitor

}
\keyword{datasets}
\keyword{ts}
