% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfast-package.R
\docType{data}
\name{simts}
\alias{simts}
\title{Simulated seasonal 16-day NDVI time series}
\source{
\insertRef{janbfast}{bfast}
}
\description{
\code{simts} is an object of class "stl" and consists of seasonal, trend
(equal to 0) and noise components. The simulated noise is typical for
remotely sensed satellite data.
}
\examples{

plot(simts)
# References
citation("bfast")

}
\keyword{datasets}
